delete app.routing.ts and paste rotes in app.module.ts
